+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master (not included in any release yet)
 * Fix TypeError when formatting logging messages that don't have any args (#15 by: DasSkelett)
+* Add 'only' parameter to Action; Changes to this repo's workflow setup (#16 by: DasSkelett)
 
 ## v1
 ## v1.3.0